+++ R/utilities.R
@@ -1,5 +1,7 @@
-
-# Copied from https://github.com/tidyverse/ggplot2/issues/2864
+#' Parse takes a vector of n lines and returns m expressions.
+#' See https://github.com/tidyverse/ggplot2/issues/2864 for discussion.
+#'
+#' @noRd
 parse_safe <- function(text) {
   stopifnot(is.character(text))
   out <- vector("expression", length(text))