+++ NEWS.md
@@ -5,6 +5,8 @@
 
 - Adds a new package website: https://const-ae.github.io/ggsignif/ (PR by
   @IndrajeetPatil).
+  
+- Minimum `ggplot2` version needed is bumped to `3.3.3`.
 
 # Version 0.6.0